+++ src/Opulence/Validation/Rules/NotInRule.php
@@ -8,11 +8,17 @@
  */
 namespace Opulence\Validation\Rules;
 
+use InvalidArgumentException;
+use LogicException;
+
 /**
  * Defines the not-in-array rule
  */
-class NotInRule extends InRule
+class NotInRule implements IRuleWithArgs
 {
+    /** @var array The value to compare against */
+    protected $array = null;
+
     /**
      * @inheritdoc
      */
@@ -26,6 +32,22 @@ class NotInRule extends InRule
      */
     public function passes($value, array $allValues = []) : bool
     {
-        return !parent::passes($value, $allValues);
+        if ($this->array === null) {
+            throw new LogicException("Array not set");
+        }
+
+        return !in_array($value, $this->array);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function setArgs(array $args)
+    {
+        if (count($args) != 1 || !is_array($args[0])) {
+            throw new InvalidArgumentException("Must pass a list of values");
+        }
+
+        $this->array = $args[0];
     }
 }